+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - true to да, false to нет in dashSettings
 - dashSVG position tooltip
 - dashMultiline position tooltip
+- dashMultiline fix highlight
 
 ## [1.56.0]